Cleanup Test Framework headers
Changes were made following clang-tidy's recommendations.
