+++ tests/framework/test_util.cpp
@@ -27,6 +27,8 @@
 
 #include "test_util.h"
 
+#include <fstream>
+
 #if defined(WIN32)
 #include <wchar.h>
 #include <strsafe.h>
