+++ tests/framework/test_util.h
@@ -44,14 +44,10 @@
 #include <algorithm>
 #include <array>
 #include <iostream>
-#include <fstream>
 #include <ostream>
 #include <string>
 #include <vector>
 #include <unordered_map>
-#include <utility>
-#include <memory>
-#include <functional>
 #include <filesystem>
 
 #include <cassert>
