+++ tests/loader_alloc_callback_tests.cpp
@@ -27,6 +27,7 @@
 
 #include "test_environment.h"
 
+#include <fstream>
 #include <mutex>
 
 struct MemoryTrackerSettings {
